refactor: restructure command execution logic
1. Introduce CommandExecutor: Moved methods directly invoked in the command execution chain from CommandHelper to CommandExecutor.

 2. Refactor LazyTickCommand: Moved indirect utility methods (executeList and executeReset) to CommandHelper. LazyTickCommand is now solely responsible for command registration.
